fix: improve chainhead unpin operations (#701)
This PR is to fix issue where `the logic watching for transaction
inclusion was throwing the unpinned block error` as reported by @bkchr
in https://github.com/dedotdev/dedot/issues/699

Logic for checking transaction inclusion is supposed to not throwing out
the `unpinned block error` as we'll need make sure to record usage for
the blocks for body checking to make sure we don't accidentally unpin
those blocks.

We do record block usage for chain head operations but those are running
in a queue which might be triggered after receiving the finalized events
which cause the blocks to be unpinned before the block-usage recording
kicks-in. We'll fix this by immediately record block usage before
register chainhead operations into the queue.

We also increase the default minimum queue size to 16 as well, a better
approach for later here is to adjust the queue size dynamically based to
the number of finalized blocks each batch.